+++ src/main/java/com/fivemybab/ittabab/user/command/application/dto/CourseDTO.java
@@ -1,4 +1,4 @@
-package com.fivemybab.ittabab.user.command.dto;
+package com.fivemybab.ittabab.user.command.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
