+++ src/main/java/com/fivemybab/ittabab/user/command/application/dto/UserDTO.java
@@ -1,4 +1,4 @@
-package com.fivemybab.ittabab.user.command.dto;
+package com.fivemybab.ittabab.user.command.application.dto;
 
 import lombok.Getter;
 import lombok.Setter;
