+++ src/main/java/com/fivemybab/ittabab/user/command/domain/repository/UserCommandRepository.java
@@ -1,4 +1,4 @@
-package com.fivemybab.ittabab.user.command.repository;
+package com.fivemybab.ittabab.user.command.domain.repository;
 
 import com.fivemybab.ittabab.user.command.domain.aggregate.UserInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
